Add enhanced movement and map navigation features
Added MovementAgent that uses A* algorithm for shortest path calculation and allows tweens to navigate through the map.
Enhanced the AimHelper with prediction of laser trajectory and player filters.
Update PythonController with the method "InputFire".
Created environment map representation with Map Graph and graph nodes, allowing path finding and environment analysis.
Updated the HumanLikeMouse controller with checks to determine if movement has ended.
Added login with the User method 'login'.
Renamed AimHelper.h to IAimHelper.h and added methods for player filtering and predicting laser shooting.
Added methods in vmath for vector comparison.
Changed the behaviour of the login menu.