Title: Re-structure API and add Vector2 functionality
Re-structured the API folder and moved api.cpp and api.h to the newly created directory. Added support for Vector2 functionality applied in the 'api' modules. This enhancement was implemented to cater for use-cases that require interacting with vectors in the game environment. The changes also include the editing of CMakeLists.txt to reflect the updated file structure.