Add missing API files and fix method declaration
In CMakeLists, three files python/api/api_collision.h, python/api/api_console.h and python/api/api_input.h were added to expand API functionality. In the API.py file, a correction in the method declaration was made in the distance function. Previously, the distance function was not properly declared, it did not include the self parameter which is necessary in Python for instance methods.