Remove warning popup for pnglite-incompatible images
Remove warning popup when loading images which are not compatible with pnglite.

The number of users of client versions older than 16.3, which is the version in which pnglite was replace with libpng, appears to be less than 2%.

Clients older than version 16.3 fail gracefully when pnglite cannot load an image.

At the same time, we should set the https://info.ddnet.org/info attribute `warn-pnglite-incompatible-images` to `false`, so the warning popup is retroactively disabled for clients with version 16.3 and newer.