Add flags for compatibility with Windows 8 - 11
We previously only defined compatibility with Windows Vista and 7. Now we define compatibility with Windows Vista - 11, which is what we currently support.

This means that Windows 8 - 11 will no longer try to run the client in a compatibility mode for Windows 7 applications. This should effectively not change anything, as we don't directly use any of the [components that function differently depending on the compatibility information](https://learn.microsoft.com/en-us/windows/compatibility/application-executable-manifest#manifestation).