Add `IJob::Done` callback called on main thread
We currently poll if specific background jobs are done in various places to perform some final computations on the main thread.

To simplify job handling the virtual function `IJob::Done` is added to perform these final computations. This function will be called automatically on the main thread by the job pool, for all jobs that have finished their `Run` function. Hence, polling the current state of jobs is not necessary anymore for most use-cases.

The `CJobPool::Update` function is added, which must be called on the main thread periodically to call the `Done` callback for all finished jobs, which are stored in a separate queue.

For the generic `CHostLookup` and `CHttpRequest` jobs the `Done` function can be set to a lambda function, so overriding these classes is not necessary just to specify a custom `Done` function.