Do not refill ammo while shooting
Holding down fire on 5 bullets gives you 6 shots.
Because it refills during the time it takes to fire.
Some gctf modes implement a timer reset on every shot.

I tested it a bit and it feels cleaner. Then the ammo is
actually empty when you shot all bullets. And there is no
random feeling chance of a bullet directly popping up again.