Add more gamestates to make code even more complex
I am tired af and will regret this decision omg

But it helps my brain right now to allow different types of countdowns.
With different delays. And one does not reset the timer but the other
one does.

Staying close to ddnet or vanilla is already a lost cause when it comes
to the timer code.