Properly uninitialize network on Windows, add RAII wrapper
On Windows, we call `WSAStartup` in `net_init`. According to the [documentation](https://learn.microsoft.com/en-us/windows/win32/api/winsock/nf-winsock-wsastartup) we should call `WSACleanup` once for every previous successful call to `WSAStartup`.

For this purpose, the function `net_uninit` is added to call `WSACleanup` on Windows. The usage of the `net_init` and `net_uninit` functions is simplified by adding a RAII wrapper for them. The network initialization in client/server is moved from the `CEngine` constructor to the respective client/server `main` functions.