Fix crashbug and revert kill message log duplication
We have to create conflict surface with ddnet here because we have
nullptr killers and ddnets CCharacter::Die() does not support that

Since we already conflict might as well get rid of the log level debug
and make it standard and then we do not have to move that logging to the
insta controller.

In ddnet there are no selfkills and in ddnet-insta it is possible to
leave the game and then have your bullets make a kill. Not sure if we
could get that if statement merged in upstream.