Do not mark cheat servers as test servers
ddnet needs to mark test servers since they do not store ranks
i would like to keep cheats on for testing and trolling
but it looks weird in the server browser

also for fast cap ideally ranks should be stored even if cheats are
allowed. But runs where super was used should not be. But I think ddnet
covers that already.

This should be revisited when fastcap is finished.

Maybe it needs a differentation between ddnet/ddrace cheats such as
super and /r and ddnet-insta cheats such as godmode.