Use temporary files for recording demos
Initially record demos to `.%pid%.tmp` files and rename them to the real filename when the recording is stopped. This ensures that `.demo` files are valid, assuming no `io_write` errors occur. If the client crashes or otherwise fails to stop the demo recording, then the invalid demos will stay `.tmp` files.

To simplify the usage, the handling of temporary files is added directly to the demo recorder. A parameter `bool RemoveFile` is added to the `IDemoRecorder::Stop` function to control whether the temporary file is removed or renamed to the real filename.

Ensure `.tmp` files of the replay recorder are always removed (except on crashes). The temporary replay demos were previously not removed when being disconnected from a server.

Ensure that replay filenames are unique by appending a random number. The replay demo is sliced in a background job, so we need to avoid recording to the same file when restarting the replay recorder.

Fix auto demos being stopped and restarted when enabling replays in the settings menu. Now only the replay recorder is stopped/started when necessary.

Fix replay recorder being restarted when setting `cl_replay` variable with console without changing the value.

Remove unnecessary `CClient::m_ButtonRender` variable. Demo rendering is already stopped correctly when reaching the end of the demo without this additional variable.

Remove unused `CDemoRecorder::m_pMapData` variable.

See #7349.